pyproject.toml
[tool.poetry]

name = "skybrush-ext-aimotionlab"
version = "1.0.0"
description = "libmotioncapture mocap support and aimotionlab extension for Skybrush Server"
authors = ["Tamas Nepusz <ntamas@gmail.com>", "Antal Péter <antalpeter@sztaki.hu>", "Gaál Botond <gaalbotond@sztaki.hu>", "Palkovits Máté <palkovitsmate@sztaki.hu>"]
license = "UNLICENSED"
packages = [
  # Change this to the name of the folder in src/ containing your extension
  { include = "skybrush_ext_libmotioncapture", from = "src" },
  { include = "skybrush_ext_aimotionlab", from = "src" }  
]

[[tool.poetry.source]]
name = "fury"
url = "https://pypi.fury.io/skybrush/"
secondary = true

[tool.poetry.dependencies]
python = "^3.8"
quart = ">=0.18.0"
motioncapture = { file = "../libmotioncapture/dist/motioncapture-1.0a1-cp310-cp310-linux_x86_64.whl"}
numpy = "^1.23.1"

[tool.poetry.dev-dependencies]
flockwave-server = { git = "https://github.com/skybrush-io/skybrush-server", branch = "main" }

[tool.poetry.plugins]

[tool.poetry.plugins."skybrush.server.ext"]
# Change the key below to the name of the extension that you want to use in
# configuration files, and the value to the name of the Python module that
# contains the extension code
libmotioncapture = "skybrush_ext_libmotioncapture"
ext_aimotionlab="skybrush_ext_aimotionlab"

[tool.poetry.scripts]
skybrushd = "flockwave.server.launcher:start"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

